+++ README.md
@@ -1,25 +1,52 @@
 ## Teste Dev Udna
 
-- [ ] 0)  Fazer um fork ou clone do projeto e rodar o projeto localmente
+- [x] 0)  Fazer um fork ou clone do projeto e rodar o projeto localmente
 
-- [ ] 1) O que você melhoraria nesse projeto? Explique brevemente como faria isso
+- [x] 1) O que você melhoraria nesse projeto? Explique brevemente como faria isso
+      Fazer o sistema mostrar, quem venceu o jogo e limpar a tela após exibir a mensagem, tal função já foi implementada no jogo.
+      Criar um banner onde o usuário poderá informa as regras do jogo, criaria uma pagina anterior a Home.
+      Bloquear usuário de continuar a jogar após, foi implementado função para reiniciar o game após mensagem de vencedor.
+      
 
-- [ ] 2) Colocar a mensagem de jogador (Vez do jogador) em cima do jogo da velha
+- [x] 2) Colocar a mensagem de jogador (Vez do jogador) em cima do jogo da velha
 
-- [ ] 3)  Muda as cores do tema dark e light (escolha uma cor qualquer diferente pra cada)
+- [x] 3)  Muda as cores do tema dark e light (escolha uma cor qualquer diferente pra cada)
 
 - [ ] 4)  Explique com suas palavras a estrutura de arquivos do projeto
+        
+        * Na pasta src está todo codigo excencial para que o game funcione.
+         
+         * na pasta assets existe as imagens utiliadas para logo do modo escuro e claro,
+         
+         * na pasta components os arquivos que podem ser usados mais de uma vez na aplicação.
+         
+          Dentro da pagina components existem:
+          Pasta Box -> é o component que redenriza o tabuleiro da DOM  e estilos
+          Pasta GameBord -> é o component que renderiza os textos de vez do jogador e seus estilos
+          Pasta header -> é o component onde é redenrizado a logo e estilos.
+          Pasta Slide -> é o component que renderiza o botão de mudança de tema e seus estilos.
+          Pasta StatusBar ->  é o component que renderiza o botão de reiniciar o jogo e seus estilos.
 
-- [ ] 5)  Explique a diferença entre os dois códigos abaixo
+          *A pasta config. está os arquivos de configuração do reactotron
+
+          *Pasta pages fica a pagina home que importa os components para o game.
+          
+          *A pasta store pelo que entendi tem haver com a logica porém não entendi ao certo seu funcionamento.
+
+          *A pasta Styles é onde foi aplicado a estilização global do game.
+
+          *A pasta utils é onde fica as regras do game onde é definida as condições de vitória.
+
+- [x] 5)  Explique a diferença entre os dois códigos abaixo
 
 ```js
   Gameboard.js
-  code: 1
+  code: 1 -> desenha um tabuleiro em branco e também altera usando um array de 9 posições.
   useEffect(()=>{
     setDraw(Array(9).fill(false))
   },[])
 
-  code: 2
+  code: 2 -> atualiza a cada rodada procurando se exite um ganhador a cada jogada.
   useEffect(() => {
     setWinner(Verifica(tabuleiro))
   }, [tabuleiro])
@@ -28,7 +55,7 @@
 
 Caso não consiga resolver alguma das tarefas explique as dificuldades encontradas
 
-- [ ] 6) Faça uma pull request nesse repositório ou mande o link com a solução 
+- [x] 6) Faça uma pull request nesse repositório ou mande o link com a solução 
 
 ---
