+++ package.json
@@ -6,13 +6,15 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "react": "^16.12.0",
+    "hooks": "^0.3.2",
+    "react": "^16.13.1",
     "react-dom": "^16.12.0",
-    "react-redux": "^7.1.3",
-    "react-scripts": "3.3.0",
+    "react-redux": "^7.2.0",
+    "react-scripts": "^3.3.1",
+    "reactotron": "^0.9.0",
     "reactotron-react-js": "^3.3.7",
     "reactotron-redux": "^3.1.2",
-    "redux": "^4.0.4",
+    "redux": "^4.0.5",
     "styled-components": "^4.4.1"
   },
   "scripts": {
