+++ src/components/Box/styles.js
@@ -9,10 +9,10 @@ export const Container = styled.div`
     height: 100px;
     width: 100px;
     background: ${props => props.tema === true ? '#e5e5e5'  : '#091b28'};
-   
+  //Cor das Letras X ou O
     :only-child {
-      color: ${props => (props.tema ? '#390047' : '#e7a2f8')};
-      font-size: 26px;
+      color: ${props => (props.tema ? '#000080' : '#ffff7f')}; 
+      font-size: 50px;
       font-weight: bold;
     }
   }
