+++ src/components/Header/styles.js
@@ -7,12 +7,14 @@ export const Container = styled.div`
   height: 50px;
   padding: 10px;
   margin-left: 100px;
-
-  h1 {
+  
+  h1 
     color: ${props => props.tema === false ? '#373c49' : '#e4e4e5'  };
     font-size: 40px;
   }
 
+  
+  
   img {
     height: 100px;
     border-radius: 5px;
